# swagger.yaml
openapi: 3.0.3
info:
  title: Redis Streams POC API
  description: |
    A comprehensive Node.js application demonstrating Redis Streams for real-time data processing with Express, PostgreSQL, and Sequelize.
    
    ## Features
    - 🚀 Redis Streams Integration with producer/consumer patterns
    - 📊 Consumer Groups for scalable and fault-tolerant message processing
    - 🗄️ PostgreSQL + Sequelize for persistent data storage
    - 🌐 RESTful API endpoints for triggering events
    - 📝 Complete audit trail of all events
    - 🔄 Real-time processing with event-driven architecture
    
    ## Redis Streams
    This API demonstrates various Redis Streams operations:
    - **XADD** - Publishing events to streams
    - **XREAD** - Simple event consumption
    - **XREADGROUP** - Consumer group processing
    - **XACK** - Message acknowledgment
    - **XGROUP CREATE** - Consumer group management
  version: 1.0.0
  contact:
    name: Redis Streams POC
    url: https://github.com/your-username/redis-streams-poc
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/events/chat:
    post:
      tags:
        - Chat
      summary: Send chat message
      description: Send a chat message and publish a message.sent event to Redis Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - senderId
                - receiverId
                - content
              properties:
                senderId:
                  type: string
                  example: user-123
                  description: ID of the message sender
                receiverId:
                  type: string
                  example: user-456
                  description: ID of the message receiver
                content:
                  type: string
                  example: Hello, how are you?
                  description: Message content
                roomId:
                  type: string
                  example: general
                  default: general
                  description: Chat room ID
      responses:
        '201':
          description: Chat message sent and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    $ref: '#/components/schemas/ChatMessage'
                  messageId:
                    type: string
                    example: "1642248000000-3"
                  message:
                    type: string
                    example: Chat message sent and event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/tasks:
    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task and publish a task.created event to Redis Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - assignedTo
                - createdBy
              properties:
                title:
                  type: string
                  example: Implement Redis Streams
                  description: Task title
                description:
                  type: string
                  example: Complete the Redis Streams POC implementation
                  description: Task description
                priority:
                  type: string
                  enum: [low, medium, high]
                  default: medium
                  description: Task priority
                assignedTo:
                  type: string
                  example: user-123
                  description: User ID assigned to the task
                createdBy:
                  type: string
                  example: user-456
                  description: User ID who created the task
                dueDate:
                  type: string
                  format: date-time
                  example: "2024-01-20T23:59:59.000Z"
                  description: Task due date
      responses:
        '201':
          description: Task created and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  task:
                    $ref: '#/components/schemas/Task'
                  messageId:
                    type: string
                    example: "1642248000000-4"
                  message:
                    type: string
                    example: Task created and event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/tasks/{taskId}/complete:
    post:
      tags:
        - Tasks
      summary: Complete task
      description: Mark a task as completed and publish a task.completed event to Redis Streams
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            example: task-123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - completedBy
              properties:
                completedBy:
                  type: string
                  example: user-123
                  description: User ID who completed the task
                completionNotes:
                  type: string
                  example: Task completed successfully with all requirements met
                  description: Optional completion notes
      responses:
        '200':
          description: Task completed and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: "1642248000000-5"
                  message:
                    type: string
                    example: Task completion event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/logs:
    post:
      tags:
        - System
      summary: Publish system log
      description: Publish a system log event to Redis Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - level
                - message
              properties:
                level:
                  type: string
                  enum: [info, warn, error, debug]
                  example: info
                  description: Log level
                message:
                  type: string
                  example: User authentication successful
                  description: Log message
                metadata:
                  type: object
                  example: { userId: "user-123", ip: "192.168.1.1" }
                  description: Additional metadata
      responses:
        '201':
          description: System log event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: "1642248000000-6"
                  message:
                    type: string
                    example: System log event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/custom:
    post:
      tags:
        - Events
      summary: Publish custom event
      description: Publish a custom event to any Redis Stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - streamName
                - eventType
                - payload
              properties:
                streamName:
                  type: string
                  example: custom:events
                  description: Name of the Redis Stream
                eventType:
                  type: string
                  example: custom.event
                  description: Type of the event
                payload:
                  type: object
                  example: { data: "custom data", timestamp: "2024-01-15T10:30:00.000Z" }
                  description: Event payload data
      responses:
        '201':
          description: Custom event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: "1642248000000-7"
                  message:
                    type: string
                    example: Custom event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/batch:
    post:
      tags:
        - Events
      summary: Batch publish events
      description: Publish multiple events to Redis Streams in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - events
              properties:
                events:
                  type: array
                  items:
                    type: object
                    required:
                      - streamName
                      - eventType
                      - payload
                    properties:
                      streamName:
                        type: string
                        example: user:events
                      eventType:
                        type: string
                        example: user.created
                      payload:
                        type: object
                        example: { userId: "user-123", username: "john_doe" }
                  example:
                    - streamName: "user:events"
                      eventType: "user.created"
                      payload: { userId: "user-123", username: "john_doe" }
                    - streamName: "chat:messages"
                      eventType: "message.sent"
                      payload: { senderId: "user-123", content: "Hello" }
      responses:
        '201':
          description: Batch events published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageIds:
                    type: array
                    items:
                      type: string
                    example: ["1642248000000-8", "1642248000000-9"]
                  count:
                    type: integer
                    example: 2
                  message:
                    type: string
                    example: Batch events published successfully
        '400':
          description: Bad request - invalid events array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/processed:
    get:
      tags:
        - Events
      summary: Get processed events
      description: Retrieve a list of processed events from the database
      parameters:
        - name: limit
          in: query
          description: Maximum number of events to return
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of events to skip
          required: false
          schema:
            type: integer
            default: 0
        - name: eventType
          in: query
          description: Filter by event type
          required: false
          schema:
            type: string
            example: user.created
        - name: processed
          in: query
          description: Filter by processed status
          required: false
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Processed events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                    example: 150
                  limit:
                    type: integer
                    example: 50
                  offset:
                    type: integer
                    example: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/stats:
    get:
      tags:
        - Events
      summary: Get event statistics
      description: Get statistics about events in the system
      responses:
        '200':
          description: Event statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 1000
                        description: Total number of events
                      processed:
                        type: integer
                        example: 950
                        description: Number of processed events
                      pending:
                        type: integer
                        example: 50
                        description: Number of pending events
                      errors:
                        type: integer
                        example: 5
                        description: Number of events with errors
                      eventTypes:
                        type: array
                        items:
                          type: object
                          properties:
                            eventType:
                              type: string
                              example: user.created
                            count:
                              type: integer
                              example: 100
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/info:
    get:
      tags:
        - Events
      summary: Get available streams and event types
      description: Get information about available Redis Streams and event types
      responses:
        '200':
          description: Stream and event type information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  streams:
                    type: array
                    items:
                      type: string
                    example: ["user:events", "chat:messages", "task:queue", "system:logs"]
                    description: Available Redis Streams
                  eventTypes:
                    type: array
                    items:
                      type: string
                    example: ["user.created", "user.updated", "user.deleted", "message.sent", "task.created", "task.completed", "system.error", "system.info"]
                    description: Available event types
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        senderId:
          type: string
          example: user-123
        receiverId:
          type: string
          example: user-456
        content:
          type: string
          example: Hello, how are you?
        roomId:
          type: string
          example: general
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: Implement Redis Streams
        description:
          type: string
          example: Complete the Redis Streams POC implementation
        priority:
          type: string
          enum: [low, medium, high]
          example: high
        assignedTo:
          type: string
          example: user-123
        createdBy:
          type: string
          example: user-456
        dueDate:
          type: string
          format: date-time
          example: "2024-01-20T23:59:59.000Z"
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        streamId:
          type: string
          example: "1642248000000-0"
          description: Redis Stream message ID
        eventType:
          type: string
          example: user.created
        payload:
          type: object
          example: { userId: "user-123", username: "john_doe" }
        processed:
          type: boolean
          example: true
        processedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        consumerId:
          type: string
          example: consumer-1
        error:
          type: string
          example: null
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: User not found
        details:
          type: string
          example: The requested user ID does not exist
        path:
          type: string
          example: /api/users/123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...